requirements.txt
matplotlib==3.4.3
numpy==1.21.5
Pillow==9.0.1
requests==2.26.0
tensorflow==2.8.0
tensorflow_gpu==2.6.0
